+++ vollseg/SmartSeeds3D.py
@@ -221,7 +221,7 @@ class SmartSeeds3D(object):
                     
                             image = imread(fname)
                             if self.erosion_iterations > 0:
-                               image = erode_labels(image, self.erosion_iterations)
+                               image = erode_labels(image.astype('uint16'), self.erosion_iterations)
                             Name = os.path.basename(os.path.splitext(fname)[0])
                     
                             Binaryimage = image > 0
